fix(sources): harden AWS account contact kernel
