+++ crates/source-runtime-next/testdata/aws_account_contact/SHA256SUMS
@@ -1,0 +1,2 @@
+6220b0a437199091e0bb3295ee92937bf06232a6991a912f33fe1268e565cfb5  primary_response.json
+200f5f6c91aa5299b473e696e2d602cdddc73bfa80683b7afaaa09a0f3a9a43d  security_response.json
