+++ crates/source-runtime-next/testdata/aws_account_contact/primary_response.json
@@ -1,0 +1,1 @@
+{"ContactInformation":{"AddressLine1":"123 Example Street","City":"Example City","CompanyName":"Example Corp","CountryCode":"US","FullName":"Private Primary","PhoneNumber":"+1-555-0100","PostalCode":"00000","StateOrRegion":"CA"}}
