+++ crates/source-runtime-next/testdata/aws_account_contact/security_response.json
@@ -1,0 +1,1 @@
+{"AlternateContact":{"AlternateContactType":"SECURITY","EmailAddress":"security@example.test","Name":"Private Security","PhoneNumber":"+1-555-0101","Title":"Security"}}
